+++ src/App.scss
@@ -6,7 +6,7 @@
 
 @import 'Containers/Home/Home.scss';
 @import 'Containers/Login/Login.scss';
-
+@import 'Containers/Registration/Registration.scss';
 
 *{
 font-family: 'Nunito';
