+++ src/Components/Header/Header.scss
@@ -8,7 +8,6 @@ padding: 1em 4em;
 display: flex;
 justify-content: space-between;
 align-items: center;
-
 }
 
 .logo{
@@ -16,15 +15,20 @@ height: 1.5em;
 cursor: pointer;
 }
 .div-print-info-correct {
-    color: rgb(63, 63, 63);
-     padding-bottom: 1em;    
-     margin-top: 0.5em;
-     height: 1em;
-     font-size: 14px;
- }
- .no-height{
-    height: unset!important;
- }
- .header-menu-links{
+color: rgb(63, 63, 63);
+padding-bottom: 1em;    
+margin-top: 0.5em;
+height: 1em;
+font-size: 14px;
+}
+.no-height{
+height: unset!important;
+}
+.header-menu-links{
 cursor: pointer;
- }
+}
+.header-menu {
+display: flex;
+width: 9em;
+justify-content: space-between;
+}
