+++ src/Containers/Registration/Registration.scss
@@ -1,0 +1,121 @@
+.registration-section-1{
+height: calc(100vh - 6.9em);
+justify-content: center;
+align-items: center;
+background: url('../../images/background-3.png');
+display: flex;
+background-color: rgb(240, 240, 240);
+width: 100%;
+height: fit-content;
+min-height: 48em;
+padding: 3em 0em;
+display: flex;
+justify-content: center;
+}
+.sendButton {
+height: 3em;
+width: 100%;
+background-color: green;
+color: white;
+border-radius: 0.2em;
+display: flex;
+margin-top: 2em;
+justify-content: center;
+align-items: center;
+font-size: 1.2em;
+cursor: pointer;
+border-radius: 30px;
+margin-bottom: 8em;
+}
+.sendButton-no-ready {
+height: 3em;
+width: 100%;
+background-color: rgb(58, 58, 58);
+color: white;
+border-radius: 0.2em;
+display: flex;
+margin-top: 2em;
+justify-content: center;
+align-items: center;
+font-size: 1.2em;
+cursor: pointer;
+border-radius: 30px;
+cursor: no-drop;
+}
+.home-input.arquitect {
+margin: 1em 0em;
+display: flex;
+justify-content: space-evenly;
+}
+#input-rol-arquitect{
+margin-left: 1em;
+}
+
+.home-form-full-w {
+width: 100%;
+}
+.home-form-fields-ok {
+color: #4c4c4c;
+line-height: 2.6em;
+font-size: 0.9em;
+display: flex;
+align-items: center;
+}
+p.home-form-budget-p {
+line-height: 0;
+font-size: 0.9em;
+}
+.home-form-city-date {
+display: flex;
+flex-direction: row;
+justify-content: space-between;
+}
+textarea.no-height.home-form-fields.no-height {
+padding: 0.5em 0em 0em 0.5em;
+}
+.home-section3-form-width{
+width: 25em;
+}
+.face_icon {
+height: 3em;
+}
+.home-section3-h4 {
+text-align: center;
+}
+.p-face-icon {
+text-align: center;
+}
+input#rol {
+    height: 1.2em;
+    width: 1.2em;
+    }
+    .home-section3-radio-text {
+        display: flex;
+    }
+    .width-100{
+    width: 100%;
+    }
+
+    .home-ask-3-architects{
+        text-align: center;
+        font-size: 1.5em;
+        font-weight: 700;
+        color: #00718a;
+        line-height: 1em;    
+        }
+        .home-form-fields-div {
+        display: flex;
+        flex-direction: column;
+        align-items: flex-start;
+        }
+        .home-form-fields {
+        margin: 0.4em 0em;
+        width: 26em;
+        height: 2.5em;
+        width: 100%;
+        border-radius: 0.2em;
+        //text-align: center;
+        padding: 0em 0em 0em 0.5em;
+        border: 1px solid grey;
+        }
+
